toolkit/components/passwordmgr/test/mochitest/test_autocomplete_new_password.html

<!DOCTYPE HTML>
<html>
<head>
  <meta charset="utf-8">
  <title>Test autofill with autocomplete=new-password fields</title>
  <script src="/tests/SimpleTest/SimpleTest.js"></script>
  <script src="/tests/SimpleTest/EventUtils.js"></script>
  <script type="text/javascript" src="pwmgr_common.js"></script>
  <script type="text/javascript" src="../../../satchel/test/satchel_common.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css" />
</head>
<body>
Login Manager test: autofill with autocomplete=new-password fields

<script>
let chromeScript = runInParent(function initLogins() {
  const {Services} = ChromeUtils.import("resource://gre/modules/Services.jsm");

  let login1 = Cc["@mozilla.org/login-manager/loginInfo;1"]
                        .createInstance(Ci.nsILoginInfo);
  login1.init("https://example.com", "https://autofill", null,
              "user1", "pass1", "", "");

  Services.logins.addLogin(login1);
});

let readyPromise = registerRunTests();
</script>
<p id="display"></p>

<!-- we presumably can't hide the content for this test. -->
<div id="content">

  <!-- form1 is the reference, sanity-check -->
  <form id="form1" action="https://autofill" onsubmit="return false;">
    <input type="text" name="uname">
    <input type="password" name="p">
    <button type="submit">Submit</button>
  </form>

  <!-- form2 uses the new-password field -->
  <form id="form2" action="https://autofill" onsubmit="return false;">
    <input type="text" name="uname">
    <input type="password" autocomplete="new-password">
    <button type="submit">Submit</button>
  </form>
</div>

<pre id="test">
<script class="testbody" type="text/javascript">
const {ContentTaskUtils} =
  SpecialPowers.Cu.import("resource://testing-common/ContentTaskUtils.jsm", {});

async function getAutocompleteResult(input, expectedValues) {
  input.focus();

  const shownPromise = promiseACShown();
  synthesizeKey("KEY_ArrowDown");
  await shownPromise;
  synthesizeKey("KEY_ArrowDown");
  await synthesizeKey("KEY_Enter");

  let didAutocomplete;
  try {
    await ContentTaskUtils.waitForCondition(() => {
      for (let [selector, expectedValue] of Object.entries(expectedValues)) {
        if (document.querySelector(selector).value !== expectedValue) {
          return false;
        }
      }
      return true;
    });
    didAutocomplete = true;
  } catch (ex) {
    info("waitForCondition exception: " + ex.message);
    didAutocomplete = false;
  }
  return didAutocomplete;
}

add_task(async function setup() {
  ok(readyPromise, "check promise is available");
  await readyPromise;
});

add_task(async function test_autofillAutocompleteNewPassword() {
  // reference form was filled as expected?
  checkForm(1, "user1", "pass1");

  // 2nd form should not be filled
  checkForm(2, "", "");

  let form = document.getElementById("form2");
  let userInput = form.querySelector("[name='uname']");

  const didAutocomplete = await getAutocompleteResult(userInput, {
    "#form2 [name='uname']": "user1",
    "#form2 [type='password']": "pass1",
  });
  ok(didAutocomplete, "Autocomplete of user and password fields should happen");
});

</script>
</pre>
</body>
</html>
